+++ examples/todos/packages/in_memory_todos_data_source/lib/src/in_memory_todos_data_source.dart
@@ -3,7 +3,7 @@ import 'package:uuid/uuid.dart';
 
 /// An in-memory implementation of the [TodosDataSource] interface.
 class InMemoryTodosDataSource implements TodosDataSource {
-  /// Map of ID -> Todo
+  /// Map of ID -> `Todo`
   final _cache = <String, Todo>{};
 
   @override