+++ packages/dart_frog/lib/src/response.dart
@@ -19,6 +19,19 @@ class Response {
           ),
         );
 
+  /// Create a [Response] with a stream of bytes.
+  Response.stream({
+    int statusCode = 200,
+    Stream<List<int>>? body,
+    Map<String, Object>? headers,
+  }) : this._(
+          shelf.Response(
+            statusCode,
+            body: body,
+            headers: headers,
+          ),
+        );
+
   /// Create a [Response] with a byte array body.
   Response.bytes({
     int statusCode = 200,
@@ -48,7 +61,7 @@ class Response {
 
   Response._(this._response);
 
-  final shelf.Response _response;
+  shelf.Response _response;
 
   /// The HTTP status code of the response.
   int get statusCode => _response.statusCode;
@@ -61,17 +74,34 @@ class Response {
   Stream<List<int>> bytes() => _response.read();
 
   /// Returns a [Future] containing the body as a [String].
-  Future<String> body() => _response.readAsString();
+  Future<String> body() async {
+    const responseBodyKey = 'dart_frog.response.body';
+    final bodyFromContext =
+        _response.context[responseBodyKey] as Completer<String>?;
+    if (bodyFromContext != null) return bodyFromContext.future;
+
+    final completer = Completer<String>();
+    try {
+      _response = _response.change(
+        context: {..._response.context, responseBodyKey: completer},
+      );
+      completer.complete(await _response.readAsString());
+    } catch (error, stackTrace) {
+      completer.completeError(error, stackTrace);
+    }
+
+    return completer.future;
+  }
 
   /// Returns a [Future] containing the form data as a [Map].
-  Future<Map<String, String>> formData() {
-    return parseFormData(headers: headers, body: body);
+  Future<FormData> formData() {
+    return parseFormData(headers: headers, body: body, bytes: bytes);
   }
 
   /// Returns a [Future] containing the body text parsed as a json object.
   /// This object could be anything that can be represented by json
   /// e.g. a map, a list, a string, a number, a bool...
-  Future<dynamic> json() async => jsonDecode(await _response.readAsString());
+  Future<dynamic> json() async => jsonDecode(await body());
 
   /// Creates a new [Response] by copying existing values and applying specified
   /// changes.