+++ packages/dart_frog/test/src/request_test.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
@@ -42,10 +43,47 @@ void main() {
       expect(request.bytes(), emits(utf8.encode(body)));
     });
 
+    test('throws exception when unable to read body', () async {
+      final exception = Exception('oops');
+      final body = Stream<Object>.error(exception);
+      final request = Request('GET', localhost, body: body);
+      expect(request.body, throwsA(exception));
+    });
+
+    test('throws exception when unable to read body multiple times', () async {
+      final exception = Exception('oops');
+      final body = Stream<Object>.error(exception);
+      final request = Request('GET', localhost, body: body);
+      expect(request.body, throwsA(exception));
+      expect(request.body, throwsA(exception));
+    });
+
     test('has correct headers', () {
       const headers = <String, String>{'foo': 'bar'};
       final request = Request('GET', localhost, headers: headers);
       expect(request.headers['foo'], equals(headers['foo']));
+    });
+
+    test('body can be read multiple times (sync)', () {
+      final body = json.encode({'test': 'body'});
+      final request = Request('GET', localhost, body: body);
+
+      expect(request.body(), completion(equals(body)));
+      expect(request.body(), completion(equals(body)));
+
+      expect(request.json(), completion(equals(json.decode(body))));
+      expect(request.json(), completion(equals(json.decode(body))));
+    });
+
+    test('body can be read multiple times (async)', () async {
+      final body = json.encode({'test': 'body'});
+      final request = Request('GET', localhost, body: body);
+
+      await expectLater(request.body(), completion(equals(body)));
+      await expectLater(request.body(), completion(equals(body)));
+
+      await expectLater(request.json(), completion(equals(json.decode(body))));
+      await expectLater(request.json(), completion(equals(json.decode(body))));
     });
 
     group('copyWith', () {