+++ packages/dart_frog_cli/lib/src/commands/dev/dev.dart
@@ -1,7 +1,6 @@
 import 'dart:collection';
 import 'dart:convert';
 import 'dart:io' as io;
-import 'dart:io';
 
 import 'package:dart_frog_cli/src/command.dart';
 import 'package:dart_frog_cli/src/commands/commands.dart';
@@ -52,7 +51,7 @@ class DevCommand extends DartFrogCommand {
   /// {@macro dev_command}
   DevCommand({
     super.logger,
-    void Function(Directory)? ensureRuntimeCompatibility,
+    void Function(io.Directory)? ensureRuntimeCompatibility,
     DirectoryWatcherBuilder? directoryWatcher,
     GeneratorBuilder? generator,
     RestorableDirectoryGeneratorTargetBuilder? generatorTarget,
@@ -79,7 +78,7 @@ class DevCommand extends DartFrogCommand {
     );
   }
 
-  final void Function(Directory) _ensureRuntimeCompatibility;
+  final void Function(io.Directory) _ensureRuntimeCompatibility;
   final DirectoryWatcherBuilder _directoryWatcher;
   final GeneratorBuilder _generator;
   final Exit _exit;