+++ packages/dart_frog_web_socket/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.0-dev.3
+
+- deps: upgrade to `Dart ">=2.19.0 <3.0.0"`
+- deps: upgrade to `very_good_analysis ^4.0.0`
+
+# 0.1.0-dev.2
+
+- fix(deps): upgrade to `shelf_web_socket ^1.0.3`
+
 # 0.1.0-dev.1
 
 - feat: initial dev release 🎉